refactor: duckdbBridge → duckdb にリネームし依存方向を修正
lib → components の依存（setWasmStatus import）を解消し、
コールバック注入パターンで components → lib の一方向に統一。
