+++ CLAUDE.md
@@ -31,7 +31,7 @@ pnpm bench:gen  # generate benchmark data
 
 - Data format conventions: see `.claude/skills/aggy-import/references/data-preparation-guide.md`
 - Component files: PascalCase `.tsx`; lib files: camelCase `.ts`
-- Module-level singleton state for infrastructure (`duckdbBridge`, `i18n`); UI state uses hooks + Context
+- Module-level singleton state for infrastructure (`duckdb`, `i18n`); UI state uses hooks + Context
 - File ordering: `imports → exports (Public API) → internal implementation`. Props interfaces stay with their exported component
-- DuckDB Wasm runs sequentially on a single Web Worker — never use `Promise.all` with `duckdbBridge` methods; call them with sequential `await`
+- DuckDB Wasm runs sequentially on a single Web Worker — never use `Promise.all` with `duckdb` methods; call them with sequential `await`
 - Prefer functional style by default. Use classes when multiple functions need shared state to avoid props drilling (e.g. `aggregateGT` — bundle related data into a class and access it via methods)
